fix: type misscasting (over int -> change Long) -Price: change PRICE constant, method return type -Winning: change yield return type

+++ src/test/java/lotto/PriceTest.java
@@ -11,11 +11,11 @@ public class PriceTest {
     void calculateTest() {
         //given
         int[] match = new int[]{0, 1, 2, 3, 4, 5, 6};
-        int expectMatch = 0 * 0 + 1 * 0 + 2 * 0
+        Long expectMatch = 0 * 0 + 1 * 0 + 2 * 0
                 + 3 * THREE_MATCH_PRICE + 4 * FOUR_MATCH_PRICE + 5 * FIVE_MATCH_PRICE + 6 * SIX_MATCH_PRICE;
 
         //when
-        int result = Price.calculate(match);
+        Long result = Price.calculate(match);
 
         //then
         assertThat(result).isEqualTo(expectMatch);
